Cleanup dependencies

+++ crates/rust_generator/tests/test.rs
@@ -1202,10 +1202,8 @@ pub fn run_pulumi_generator_test(
         root.join("rust-toolchain.toml"),
     )?;
 
-    if std::env::var_os("REMOVE_LOCK_FILES").is_some() {
-        if root.join("Cargo.lock").exists() {
-            fs::remove_file(root.join("Cargo.lock"))?;
-        }
+    if std::env::var_os("REMOVE_LOCK_FILES").is_some() && root.join("Cargo.lock").exists() {
+        fs::remove_file(root.join("Cargo.lock"))?;
     }
 
     if let Some(env) = std::env::var_os("DO_NOT_COMPILE") {
